Clang complaining about types
1- Casting the float to integer because clang was complaining about
   the type
2- For some reason, the UUID in my machine is longer than the one from
   the current test. I just add another check to validate it according
   the string size.

+++ types/testing/date_test.cc
@@ -119,7 +119,7 @@ BOOST_AUTO_TEST_CASE( test_print_format )
 
 BOOST_AUTO_TEST_CASE( test_utc_parse )
 {
-    Date d(2012, 06, 06, 15, 15, 38.380);
+    Date d(2012, 06, 06, 15, 15, static_cast<int>(38.380));
     string s = "2012-Jun-06 15:15:38.380";
     Date d2 = Date::parseDefaultUtc(s);
     BOOST_CHECK_EQUAL(d2.print(3), s);
